% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/co2_20ma.R
\name{co2_20ma}
\alias{co2_20ma}
\title{20 Million Year Atmospheric CO2 Reconstruction}
\format{A \code{\link[uts:uts]{uts}} object.}
\source{
This dataset was imported by \code{\link{download_co2_20ma}} from a website by the \href{https://www.ncdc.noaa.gov/paleo/study/10423}{NOAA Earth System Research Laboratory}.
}
\description{
A 20 million year (ma) reconstruction of the atmposheric CO2 concentration (in parts per million).
}
\examples{
data(co2_20ma)
plot(co2_20ma, max_dt=dyears(1e6), type="o")      # connect observations less than 1 ma apart
plot(tail_t(co2_20ma, dyears(1.5e6)),  type="o")  # plot the 1.5 million most recent years
}
\seealso{
\code{\link{download_co2_20ma}}

\code{\link{co2_ml}} for monthly data since 1958.

\code{\link[datasets:co2]{co2}} in base \R.
}
\keyword{datasets}

