% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_ml.R
\name{co2_ml}
\alias{co2_ml}
\title{Mauna Loa Atmospheric CO2 Concentration}
\format{A \code{\link[uts:uts]{uts}} object.}
\source{
This dataset was imported by \code{\link{download_co2_ml}} in December 2015 from a website by the \href{http://www.esrl.noaa.gov/gmd/ccgg/trends/}{NOAA Earth System Research Laboratory}.

Dr. Pieter Tans, NOAA/ESRL (www.esrl.noaa.gov/gmd/ccgg/trends/) and Dr. Ralph Keeling, Scripps Institution of Oceanography (scrippsco2.ucsd.edu/).
}
\description{
The monthly average atmospheric CO2 concentration (in parts per million), as measured at the \href{http://www.esrl.noaa.gov/gmd/obop/mlo/}{Mauna Loa observatory}.
}
\examples{
data(co2_ml)
plot(co2_ml)

# Most consecutive observations are one month apart
table(round(diff(co2_ml$times) * 12 / 365))
}
\seealso{
\code{\link{download_co2_ml}} gets the most recent data.

\code{\link{co2_20ma}} for a 20 million year reconstruction.

The \code{\link[datasets:co2]{co2}} dataset in base \R is very similar, but ends in 1997 and has several missing values filled in using linear interpolation.
}
\keyword{datasets}

