% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paris_precipitation.R
\name{dowload_paris_precipitation}
\alias{dowload_paris_precipitation}
\title{Paris Monthly Precipitation 1688-2009}
\usage{
dowload_paris_precipitation()
}
\description{
This function downloads the monthly precipitation (in mm) in Paris from 1688 to 2009 from a website by the \href{https://www.ncdc.noaa.gov/paleo/study/8761}{NOAA National Climatic Data Center}. The downloaded data is subsequently imported into \R and returned as a \code{\link{uts}} object.
}
\details{
Users without internet connection can access the already imported data using \code{data(paris_precipitation)}.
}
\examples{
paris_precipitation <- dowload_paris_precipitation()
plot(paris_precipitation, max_dt=dyears(1), type="o", cex=0.5)

# Most consecutive observations are one month apart
table(round(diff(time(paris_precipitation)) / 365 * 12))

# Save data
\dontrun{
  save(paris_precipitation, file=file.path("data", "paris_precipitation.rda"), compress="xz")
}
}
\keyword{datasets}
\keyword{internal}

