% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CO2.R
\name{download_CO2}
\alias{download_CO2}
\title{Mauna Loa Atmospheric CO2 Concentration}
\usage{
download_CO2()
}
\description{
This function downloads the monthly average atmospheric CO2 concentration (in parts per million), as measured at the Mauna Loa observatory, from a website by the \href{http://www.esrl.noaa.gov/gmd/ccgg/trends/}{NOAA Earth System Research Laboratory}. The downloaded data is subsequently imported into \R and returned as a \code{\link{uts}} object.
}
\examples{
CO2 <- download_CO2()
plot(CO2)

# Most consecutive observations are one month apart
table(diff(time(CO2)))
}
\seealso{
The \code{\link[datasets:co2]{co2}} dataset in base \R is very similar, but ends in 1997 and has several missing values filled in using linear interpolation.
}
\keyword{datasets}

