% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CO2_20ma.R
\name{download_CO2_20ma}
\alias{download_CO2_20ma}
\title{20 Million Year Atmospheric CO2 Reconstruction}
\usage{
download_CO2_20ma()
}
\description{
This function downloads a 20 million year (ma) reconstruction of the atmposheric CO2 concentration (in parts per million) from a website by the \href{https://www.ncdc.noaa.gov/paleo/study/10423}{NOAA Earth System Research Laboratory}. The downloaded data is subsequently imported into \R and returned as a \code{\link{uts}} object.
}
\details{
Users without internet connection can access the already imported data using \code{data(CO2_20ma)}.
}
\examples{
CO2_20ma <- download_CO2_20ma()
plot(CO2_20ma, max_dt=dyears(1e6), type="o")      # connect observations less than 1 ma apart
plot(tail_t(CO2_20ma, dyears(1.5e6)),  type="o")  # plot the 1.5 million most recent years

# Save data
\dontrun{
  save(CO2_20ma, file=file.path("data", "CO2_20ma.rda"), compress="xz")
}
}
\seealso{
\code{\link[datasets:co2]{co2}}, \code{\link{download_CO2}}
}
\keyword{datasets}
\keyword{internal}

