% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_ml.R
\name{download_co2_ml}
\alias{download_co2_ml}
\title{Mauna Loa Atmospheric CO2 Concentration (Latest Data)}
\usage{
download_co2_ml()
}
\description{
This function downloads the monthly average atmospheric CO2 concentration (in parts per million), as measured at the \href{http://www.esrl.noaa.gov/gmd/obop/mlo/}{Mauna Loa observatory}, from a website by the \href{http://www.esrl.noaa.gov/gmd/ccgg/trends/}{NOAA Earth System Research Laboratory}. The downloaded data is subsequently imported into \R and returned as a \code{\link[uts]{uts}} object.
}
\examples{
co2_ml <- download_co2_ml()
plot(co2_ml)
table(round(diff(co2_ml$times) * 12 / 365))  # most observations are one month apart

# Save data
\dontrun{
  save(co2_ml, file=file.path("data", "co2_ml.rda"), compress="xz")
}
}
\seealso{
Users without internet connection can access the already imported data via \code{\link{co2_ml}}. However, to get the most recent data it is recommended to use this function.

\code{\link{co2_20ma}} for a 20 million year reconstruction.

The \code{\link{co2}} dataset in base \R is very similar, but ends in 1997 and has several missing values filled in using linear interpolation.
}
\keyword{datasets}
