% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes.R
\name{download_grapes}
\alias{download_grapes}
\title{Grape Harvest Dates}
\usage{
download_grapes()
}
\description{
This function creates the \code{\link{grapes}} dataset by downloading the data from a website by the \href{https://www.ncdc.noaa.gov/cdo/f?p=519:1:0::::P1_STUDY_ID:13194}{NOAA National Climatic Data Center} and returning it as a \code{\link[utsMultivariate:uts_vector]{uts_vector}} object. It is not meant to be called directly, but provided for reproducability.
}
\details{
The raw data is available in Excel and raw text, but the latter has several formatting errors. Therefore, the data is imported from the Excel file using the \href{https://cran.r-project.org/web/packages/XLConnect/index.html}{XLConnect} package.
}
\examples{
if (requireNamespace("XLConnect", quietly=TRUE)) {
  grapes <- download_grapes()
  
  \dontrun{
    save(grapes, file=file.path("data", "grapes.rda"), compress="xz")
  }
}
}
\keyword{datasets}
\keyword{internal}
