% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paris_precipitation.R
\name{download_paris_precipitation}
\alias{download_paris_precipitation}
\title{Paris Monthly Precipitation 1688-2009}
\usage{
download_paris_precipitation()
}
\description{
This function creates the \code{\link{paris_precipitation}} dataset by downloading the data from a website by the \href{https://www.ncdc.noaa.gov/paleo/study/8761}{NOAA National Climatic Data Center} and returning it as a \code{\link[uts]{uts}} object. It is not meant to be called directly, but provided for reproducability.
}
\examples{
paris_precipitation <- download_paris_precipitation()

# Save data
\dontrun{
  save(paris_precipitation, file=file.path("data", "paris_precipitation.rda"), compress="xz")
}
}
\keyword{datasets}
\keyword{internal}
