% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunspots.R
\name{download_sunspots_daily}
\alias{download_sunspots_daily}
\title{Daily Sunspot Numbers (Latest Data)}
\usage{
download_sunspots_daily()
}
\description{
Download the daily number of sunspots since 1818 from \url{http://www.sidc.be/silso/datafiles}. The data is subsequently imported into \R and returned as a \code{\link[utsMultivariate]{uts_vector}} object.
}
\examples{
\dontrun{
  # Download and save the data
  sunspots_daily <- download_sunspots_daily()
  save(sunspots_daily, file=file.path("data", "sunspots_daily.rda"), compress="xz")
}
}
\seealso{
Users without internet connection can access the already imported data via \code{\link{sunspots_daily}}. However, to get the most recent data it is recommended to use this function.
}
