% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grapes.R
\name{grapes}
\alias{grapes}
\title{Grape Harvest Dates}
\format{A \code{\link[utsMultivariate]{uts_vector}} object. Harvest dates are presented as the number of days after 31 August.}
\source{
This dataset was imported by \code{\link{download_grapes}} in June 2018 from a website by the \href{https://www.ncdc.noaa.gov/cdo/f?p=519:1:0::::P1_STUDY_ID:13194}{NOAA National Climatic Data Center}.

Daux, V., I. Garcia de Cortazar-Atauri, P. Yiou, I. Chuine, E. Garnier, E. Le Roy Ladurie, O. Mestre, and J. Tardaguila. 2011. An open-database of Grape Harvest dates for climate research: data description and quality assessment. Climate of the Past, Vol. 8, pp. 1403-1418, 2012 www.clim-past.net/8/1403/2012/ doi:10.5194/cp-8-1403-2012
}
\description{
A 650-year history of grape harvest dates for 27 regions in Western Europe.
}
\examples{
data(grapes)
plot(grapes$Burgundy)

# Connect observations less than two years apart by a line
plot(grapes$Bordeaux, max_dt=dyears(2), type="o", cex=0.5)

# Plot 20-year two-sided rolling average of Burgundy harvest dates
if (requireNamespace("utsOperators", quietly=TRUE)) {
  plot(utsOperators::rolling_apply(grapes$Burgundy, width=dyears(20), FUN=mean, align="center"))
}
}
\keyword{datasets}
