% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lynx.R
\name{import_lynx_1942}
\alias{import_lynx_1942}
\title{Annual Canadian Lynx trappings 1821-1939}
\usage{
import_lynx_1942()
}
\value{
A \code{\link[utsMultivariate:uts_vector]{uts_vector}} object. An observation for date, let's say, 1840-05-31 represents the total number of lynx catches for outfit 1839, which ran from 1 June 1839 to 31 May 1840.
}
\description{
The annual number of Canadian lynx (\emph{Lynx canadensis}) catches by the Hudson's Bay Company. Each recorded year (called \emph{Outfit}) counts collected furs from 1 June of that year until 31 May of the following year.
}
\details{
This function imports the lynx data from a local CSV file and returns it as as \code{\link{uts_vector}} object.
}
\examples{
lynx_1942 <- import_lynx_1942()
plot(lynx_1942[["MacKenzie River"]])
#plot(lynx_1942, max_dt=dyears(2), type="o", cex=0.5)    # not implemented yet

# Save data
\dontrun{
  save(lynx_1942, file=file.path("data", "lynx_1942.rda"), compress="xz")
}
}
\references{
Elton, C. and Nicholson, M. (1942). \href{http://www.jstor.org/stable/1358}{The Ten-Year Cycle in Numbers of the Lynx in Canada}. \emph{Journal of Animal Ecology}, Vol 11, No. 2, pp. 215-244
}
\seealso{
The already imported data can be accessed via \code{\link{lynx_1942}}.

The \code{\link[datasets:lynx]{lynx}} dataset in base \R is similar, except that it (i) contains data for only a single geographic region (instead of all eleven regions), (ii) has an obvious typo (compared to its quoted source).
}
\keyword{datasets}
\keyword{internal}

