% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paris_precipitation.R
\name{paris_precipitation}
\alias{paris_precipitation}
\title{Paris Monthly Precipitation 1688-2009}
\format{A \code{\link[uts:uts]{uts}} object.}
\source{
This dataset was imported by \code{\link{download_paris_precipitation}} in December 2015 from a website by the \href{https://www.ncdc.noaa.gov/paleo/study/8761}{NOAA National Climatic Data Center}.

Slonosky, V.C. 2002. Wet winters, dry summers? Three centuries of precipitation data from Paris Geophys. Res. Lett., Vol. 29(19), 1895
}
\description{
The monthly precipitation (in mm) in Paris from 1688 to 2009.
}
\examples{
data(paris_precipitation)
plot(paris_precipitation, max_dt=ddays(31), type="o", cex=0.25)

# Most consecutive observations are one month apart
table(round(diff(time(paris_precipitation)) / 365 * 12))

# Plot 20-year two-sided rolling average
if (requireNamespace("utsOperators", quietly=TRUE)) {
  roll_avg <- utsOperators::rolling_apply(paris_precipitation, width=dyears(20), FUN=mean,
    align="center", interior=TRUE)
  plot(roll_avg, max_dt=dyears(1))
}
}
\keyword{datasets}

