% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{utsData-internal}
\alias{utsData-internal}
\title{Internal Functions}
\description{
The following internal functions allow to reproduce the data download and import of the datasets provided by this package:
\itemize{
  \item \code{\link{download_co2_20ma}}
  \item \code{\link{download_grapes}}
  \item \code{\link{download_paris_precipitation}}
  \item \code{\link{import_lynx_1942}}
}
}

