% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/paris_precipitation.R
\name{dowload_paris_precipitation}
\alias{dowload_paris_precipitation}
\title{Paris Monthly Precipitation 1688-2009}
\usage{
dowload_paris_precipitation()
}
\description{
This function downloads the monthly precipitation (in mm) in Paris from 1688 to 2009 from a website by the \href{https://www.ncdc.noaa.gov/paleo/study/8761}{NOAA National Centers for Environmental Information}. The downloaded data is subsequently imported into \R and returned as a \code{\link{uts}} object.
}
\examples{
rain <- dowload_paris_precipitation()
plot(rain, max_dt=dyears(1), type="o", cex=0.5)

# Most consecutive observations are one month apart
table(round(diff(time(rain)) / 365 * 12))
}
\keyword{datasets}

