% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/co2_20ma.R
\name{download_co2_20ma}
\alias{download_co2_20ma}
\title{20 Million Year Atmospheric CO2 Reconstruction}
\usage{
download_co2_20ma()
}
\description{
This function downloads a 20 million year (ma) reconstruction of the atmposheric CO2 concentration (in parts per million) from a website by the \href{https://www.ncdc.noaa.gov/paleo/study/10423}{NOAA Earth System Research Laboratory}. The downloaded data is subsequently imported into \R and returned as a \code{\link[uts:uts]{uts}} object.
}
\examples{
co2_20ma <- download_co2_20ma()
plot(co2_20ma, max_dt=dyears(1e6), type="o")      # connect observations less than 1 ma apart
plot(tail_t(co2_20ma, dyears(1.5e6)),  type="o")  # plot the 1.5 million most recent years

# Save data
\dontrun{
  save(co2_20ma, file=file.path("data", "co2_20ma.rda"), compress="xz")
}
}
\seealso{
Users without internet connection can access the already imported data via \code{\link{co2_20ma}}.

\code{\link{co2_ml}} for monthly data since 1958.

\code{\link[datasets:co2]{co2}} in base \R.
}
\keyword{datasets}
\keyword{internal}

