% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lynx.R
\name{import_lynx_1942}
\alias{import_lynx_1942}
\title{Annual Canadian Lynx trappings 1821-1939}
\usage{
import_lynx_1942()
}
\description{
This function generates the \code{\link{lynx_1942}} data by importing the observations from a local CSV file and returning them as as \code{\link{uts_vector}} object. It is not meant to be called directly, but provided for reproducability.
}
\examples{
# Show the raw data
file.show(system.file(file.path("exdata", "lynx_1942.csv"), package="utsData"))

# Import and save the data
lynx_1942 <- import_lynx_1942()
\dontrun{
  save(lynx_1942, file=file.path("data", "lynx_1942.rda"), compress="xz")
}
}
\keyword{datasets}
\keyword{internal}

