% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{utsData-internal}
\alias{utsData-internal}
\title{Internal Functions}
\description{
The internal functions listed below allow to reproduce the data download and import of the datasets provided by this package.
}
\details{
\itemize{
  \item \code{\link{download_CO2_20ma}}
  \item \code{\link{dowload_paris_precipitation}}
  \item \code{\link{download_grapes}}
  \item \code{\link{import_lynx}}
}
}

